Set up unit tests for location and region logic
In order to keep the tests performant the Location class had to be given a method to only use the access logic, without involving the region logic. Without this the test fixture has to combine and run all logical permutations, which kill the execution time.

Logic involving prizes (pendants and crystals) is simplified by removing those prizes and relying on the "all" check always being true for no elements in `CanAquireAll`, and changing the check in `CanAquire` to be true when there is no reward.